+++ activerecord/test/models/zine.rb
@@ -1,3 +1,7 @@
 class Zine < ActiveRecord::Base
   has_many :interests, :inverse_of => :zine
+
+  def validate_i_am_fun
+    !interests.to_a.any?{|i| i.topic =~ /boringz/}
+  end
 end